Beginning the rewrite
after a bit of thinking there are a few things that i wish where
different enough that a pretty big rewrite of most of the
programme/modules was necessary. since the programme is pretty small a
total rewrite doesn't seem to bad since there wouldn't be much that
didnt get rewritten including tests.